fix: mirror virtual ports with their respective rpcport to ensure cross app compatibility
Generally apps will not recognize what port 1309 is for. However if it includes 8332 most wallet apps will recognize it as Bitcoin Core mainnet. This improves UX for standup users across existing and future external apps. It is also poor practice to use the same virtual port for multiple hidden services and is not recommended by the tor project.